refactor(console): extract first-chat handoff ownership
