+++ tldw_chatbook/UI/Console_Modules/session.py
@@ -52,7 +52,9 @@ screen-calls-its-own-controller traffic, unchanged).
 
 Moved: every `ChatScreen` method matching `*session*` whose body touched
 only session-lifecycle state (or one of the callables above) and no DOM --
-27 methods, plus four small module-level pure helpers only those methods (or
+27 methods, plus the first-chat handoff family (eight methods and its
+notification-revision state), and four small module-level pure helpers only
+those methods (or
 this cluster's own remaining screen-side callers) use:
 `_has_selected_text`/`_is_empty_select_value` (Select-value predicates
 `_default_console_session_settings` needs; `ChatScreen` imports both back --
@@ -172,6 +174,7 @@ from ...Chat.console_project_instructions import (
 )
 from ...Chat.console_session_settings import (
     ConsoleSessionSettings,
+    build_default_console_session_settings,
     build_console_settings_readiness,
     default_console_session_settings,
 )
@@ -188,7 +191,16 @@ from ...Character_Chat.persona_visual_identity import (
     resolve_persona_visual_identity,
 )
 from ...DB.VisualIdentity_DB import VisualIdentityRepository
-from ...config import coerce_bool_setting
+from ...config import (
+    coerce_bool_setting,
+    get_runtime_config_snapshot,
+    run_if_runtime_config_generation_current,
+)
+from ..Navigation.pending_handoff_store import (
+    ConsoleFirstChatIntent,
+    HandoffChannel,
+    PendingHandoffStore,
+)
 from ...Widgets.Console import (
     ConsoleComposerUndoHistory,
     ConsoleProjectInstructionStatusRow,
@@ -570,6 +582,10 @@ class ConsoleSessionController:
             [], ConsoleReactionPreviewCoordinator
         ],
         refresh_character_avatar: Callable[..., Any],
+        screen_mounted_accessor: Callable[[], bool],
+        first_chat_presentation_snapshot: Callable[[], tuple[Any, Any, object | None]],
+        apply_first_chat_control_selection: Callable[[Any, Any], None],
+        restore_first_chat_focus: Callable[[object | None], None],
     ) -> None:
         """Build the controller and bind everything its moved bodies need.
 
@@ -579,6 +595,13 @@ class ConsoleSessionController:
         "Controller-to-controller seam" section), which drop their
         `_workspace.` prefix in favour of a same-named property on this
         controller instead.
+
+        The eight-method first-chat family is likewise copied with only its
+        presentation edges adapted to the four exact late-bound callbacks:
+        mounted state, provider/model/focus snapshot, control selection, and
+        focus restoration. It retains the documented legacy `_screen`
+        exception only for the framework services above; the moved family
+        itself never reaches through `_screen` or into the DOM.
 
         Args:
             screen: The Console screen. Used ONLY for the framework
@@ -684,6 +707,13 @@ class ConsoleSessionController:
                 Console screens cannot escape an older screen's draining work.
             refresh_character_avatar: Late-bound forced avatar refresh after
                 a validated manual reaction change.
+            screen_mounted_accessor: Late-bound presentation-only mounted state.
+            first_chat_presentation_snapshot: Late-bound provider/model/focus
+                snapshot used by first-chat rollback.
+            apply_first_chat_control_selection: Late-bound projection of the
+                first-chat provider/model selection onto screen controls.
+            restore_first_chat_focus: Late-bound restoration of an opaque focus
+                token after the native async projection is synchronized.
         """
         self._screen = screen
         self.app_instance = app_instance
@@ -721,6 +751,10 @@ class ConsoleSessionController:
             reaction_preview_coordinator_accessor
         )
         self._refresh_character_avatar_fn = refresh_character_avatar
+        self._screen_mounted_accessor = screen_mounted_accessor
+        self._first_chat_presentation_snapshot_fn = first_chat_presentation_snapshot
+        self._apply_first_chat_control_selection_fn = apply_first_chat_control_selection
+        self._restore_first_chat_focus_fn = restore_first_chat_focus
 
         # This cluster's own state, moved verbatim from `ChatScreen.__init__`.
         self._console_visible_draft_session_id: str | None = None
@@ -742,6 +776,7 @@ class ConsoleSessionController:
         self._console_project_instruction_refresh_completed: dict[
             str, tuple[tuple[Any, Any], float]
         ] = {}
+        self._first_chat_handoff_notified_revision: int | None = None
 
     # -- Framework services (live-read via `@property`) --------------------
 
@@ -898,6 +933,349 @@ class ConsoleSessionController:
         """`ConsoleWorkspaceController._console_session_id_for_workspace_
         conversation`. Same prefix-drop as above."""
         return self._session_id_for_workspace_conversation_fn
+
+    @staticmethod
+    def _first_chat_defaults_match(
+        intent: ConsoleFirstChatIntent,
+        settings: ConsoleSessionSettings,
+    ) -> bool:
+        return (
+            provider_config_key(settings.provider) == intent.provider
+            and str(settings.model or "").strip() == intent.model
+        )
+
+    def _current_first_chat_defaults(
+        self,
+        *,
+        provider: str,
+        model: str,
+        config_revision: int,
+    ) -> ConsoleSessionSettings | None:
+        """Resolve exact current defaults only while the config fence matches."""
+
+        snapshot = get_runtime_config_snapshot()
+        if snapshot.generation != config_revision:
+            return None
+        settings = build_default_console_session_settings(snapshot.values)
+        if (
+            provider_config_key(settings.provider) != provider_config_key(provider)
+            or str(settings.model or "").strip() != str(model or "").strip()
+        ):
+            return None
+        return settings
+
+    def eligible_console_first_chat_session_id(self) -> str | None:
+        """Return an exact untouched target without creating or changing Console."""
+
+        store = self._console_chat_store
+        if store is None:
+            return None
+        active_id = store.active_session_id
+        if active_id is None:
+            return None
+        active = next(
+            (session for session in store.sessions() if session.id == active_id),
+            None,
+        )
+        baseline = active.canonical_settings_baseline if active is not None else None
+        if (
+            baseline is None
+            or active.workspace_id != CONSOLE_GLOBAL_WORKSPACE_ID
+            or not store.is_pristine_session(
+                active_id,
+                expected_settings=baseline,
+            )
+        ):
+            return None
+        return active_id
+
+    def _release_first_chat_claim(self, claim, message: str) -> bool:
+        """Release an exact claim without leaking failure-owned data."""
+
+        handoffs = self.app_instance.pending_handoffs
+        try:
+            claim_is_current = handoffs.is_current_claim(claim)
+        except Exception as exc:  # noqa: BLE001 - lifecycle boundary containment
+            claim_is_current = False
+            self._log_first_chat_handoff_exception("claim-current-check", exc)
+        try:
+            released = handoffs.release(claim)
+        except Exception as exc:  # noqa: BLE001 - keep the channel retryable
+            self._log_first_chat_handoff_exception("claim-release", exc)
+            released = False
+            if isinstance(handoffs, PendingHandoffStore):
+                try:
+                    # Bypass a failing instance wrapper while retaining the
+                    # store's exact-claim and replacement invariants.
+                    released = PendingHandoffStore.release(handoffs, claim)
+                except Exception as fallback_exc:  # noqa: BLE001
+                    self._log_first_chat_handoff_exception(
+                        "claim-release-fallback",
+                        fallback_exc,
+                    )
+        if not released:
+            return False
+        if not claim_is_current:
+            if self._first_chat_handoff_notified_revision == claim.revision:
+                self._first_chat_handoff_notified_revision = None
+            return False
+        if claim.revision != self._first_chat_handoff_notified_revision:
+            self._first_chat_handoff_notified_revision = claim.revision
+            try:
+                self.app_instance.notify(message, severity="warning")
+            except Exception as exc:  # noqa: BLE001 - lifecycle boundary containment
+                self._log_first_chat_handoff_exception("notification", exc)
+        return False
+
+    @staticmethod
+    def _log_first_chat_handoff_exception(category: str, exc: Exception) -> None:
+        """Log only allowlisted failure classification, never exception content."""
+
+        logger.warning(
+            "First-chat handoff operation failed (category={}, error_type={})",
+            category,
+            type(exc).__name__,
+        )
+
+    async def _resync_console_after_first_chat_rollback(
+        self,
+        prior_focused_widget: object | None,
+    ) -> None:
+        """Re-render restored Console state, then restore still-mounted focus."""
+
+        if not self._screen_mounted_accessor():
+            return
+        await self._sync_native_console_chat_ui()
+        if not self._screen_mounted_accessor():
+            return
+        self._restore_first_chat_focus_fn(prior_focused_widget)
+
+    def _resync_mounted_console_after_first_chat_rollback(
+        self,
+        *,
+        prior_control_provider: str | None,
+        prior_control_model: str | None,
+        prior_focused_widget: object | None,
+    ) -> None:
+        """Restore first-chat-owned scalars and every mounted Console projection."""
+
+        self._apply_first_chat_control_selection_fn(
+            prior_control_provider,
+            prior_control_model,
+        )
+        if not self._screen_mounted_accessor():
+            return
+        self._sync_console_chat_core_state()
+        self._sync_console_settings_summary()
+        self._sync_console_control_bar()
+        self.run_worker(
+            self._resync_console_after_first_chat_rollback(prior_focused_widget),
+            group="console-first-chat-rollback",
+            exit_on_error=False,
+        )
+
+    def consume_pending_console_first_chat_intent(self) -> bool:
+        """Activate one exact first-run target without overwriting user state."""
+
+        claim = self.app_instance.pending_handoffs.claim(
+            HandoffChannel.CONSOLE_FIRST_CHAT
+        )
+        if claim is None:
+            return False
+        intent = claim.value
+        if not isinstance(intent, ConsoleFirstChatIntent):
+            return self._release_first_chat_claim(
+                claim,
+                "The first chat could not be opened yet; review provider setup.",
+            )
+        defaults = self._current_first_chat_defaults(
+            provider=intent.provider,
+            model=intent.model,
+            config_revision=intent.config_revision,
+        )
+        if defaults is None:
+            return self._release_first_chat_claim(
+                claim,
+                "Provider settings changed before Console opened. Review setup and try again.",
+            )
+
+        store = self._ensure_console_chat_store()
+        prior_active_id = store.active_session_id
+        (
+            prior_control_provider,
+            prior_control_model,
+            prior_focused_widget,
+        ) = self._first_chat_presentation_snapshot_fn()
+        created_target = None
+        refreshed_prior: (
+            tuple[
+                ConsoleSessionSettings,
+                ConsoleSessionSettings,
+                str,
+            ]
+            | None
+        ) = None
+
+        def rollback_mutation() -> None:
+            if created_target is not None:
+                store.rollback_created_pristine_session(
+                    created_target.id,
+                    expected_session=created_target,
+                    expected_settings=defaults,
+                    prior_active_session_id=prior_active_id,
+                )
+            elif refreshed_prior is not None:
+                prior_settings, prior_baseline, prior_updated_at = refreshed_prior
+                store.rollback_pristine_session_refresh(
+                    intent.session_id,
+                    expected_current_settings=defaults,
+                    prior_settings=prior_settings,
+                    prior_canonical_settings=prior_baseline,
+                    prior_updated_at=prior_updated_at,
+                )
+            self._resync_mounted_console_after_first_chat_rollback(
+                prior_control_provider=prior_control_provider,
+                prior_control_model=prior_control_model,
+                prior_focused_widget=prior_focused_widget,
+            )
+
+        def rollback_and_release(message: str) -> bool:
+            try:
+                rollback_mutation()
+            except Exception as exc:  # noqa: BLE001 - lifecycle boundary containment
+                self._log_first_chat_handoff_exception("rollback", exc)
+            return self._release_first_chat_claim(claim, message)
+
+        def fence_matches(*, expected_active_id: str) -> bool:
+            current = self._current_first_chat_defaults(
+                provider=intent.provider,
+                model=intent.model,
+                config_revision=intent.config_revision,
+            )
+            return (
+                current == defaults
+                and store.active_session_id == expected_active_id
+                and self.app_instance.pending_handoffs.is_current_claim(claim)
+            )
+
+        reserves_new_target = (
+            self.app_instance.pending_handoffs.claim_reserves_new_console_session(claim)
+        )
+        target = next(
+            (
+                session
+                for session in store.sessions()
+                if session.id == intent.session_id
+            ),
+            None,
+        )
+        if target is None:
+            if not reserves_new_target:
+                return self._release_first_chat_claim(
+                    claim,
+                    "The intended Console session is no longer available. Review setup and try again.",
+                )
+            try:
+                target = store.create_session(
+                    session_id=intent.session_id,
+                    workspace_id=CONSOLE_GLOBAL_WORKSPACE_ID,
+                    settings=defaults,
+                    canonical_settings_baseline=defaults,
+                    activate=False,
+                )
+            except ValueError:
+                return self._release_first_chat_claim(
+                    claim,
+                    "The intended Console session was claimed before setup finished. It was left unchanged.",
+                )
+            created_target = target
+            if not fence_matches(expected_active_id=prior_active_id):
+                return rollback_and_release(
+                    "Provider settings changed while Console prepared the first chat. It will retry.",
+                )
+            store.switch_session(intent.session_id)
+            if not fence_matches(expected_active_id=intent.session_id):
+                return rollback_and_release(
+                    "Console changed while the first chat was opening. Your sessions were left unchanged.",
+                )
+        else:
+            if reserves_new_target:
+                return self._release_first_chat_claim(
+                    claim,
+                    "The intended Console session was claimed before setup finished. It was left unchanged.",
+                )
+            if store.active_session_id != intent.session_id:
+                return self._release_first_chat_claim(
+                    claim,
+                    "Console changed sessions before setup finished. Your current session was left unchanged.",
+                )
+            baseline = target.canonical_settings_baseline
+            if (
+                baseline is None
+                or target.workspace_id != CONSOLE_GLOBAL_WORKSPACE_ID
+                or not store.is_pristine_session(
+                    intent.session_id,
+                    expected_settings=baseline,
+                )
+            ):
+                return self._release_first_chat_claim(
+                    claim,
+                    "The intended Console session now contains work. It was left unchanged.",
+                )
+            if baseline != defaults:
+                refreshed_prior = (target.settings, baseline, target.updated_at)
+                store.refresh_pristine_session_settings(
+                    intent.session_id,
+                    prior_canonical_settings=baseline,
+                    current_canonical_settings=defaults,
+                )
+                target = next(
+                    session
+                    for session in store.sessions()
+                    if session.id == intent.session_id
+                )
+                if not fence_matches(expected_active_id=intent.session_id):
+                    return rollback_and_release(
+                        "Provider settings changed while Console prepared the first chat. It will retry.",
+                    )
+
+        if (
+            target.settings is None
+            or not self._first_chat_defaults_match(intent, target.settings)
+            or not fence_matches(expected_active_id=intent.session_id)
+        ):
+            return rollback_and_release(
+                "The first chat target no longer matches provider setup. It was left unchanged.",
+            )
+
+        self._apply_first_chat_control_selection_fn(
+            target.settings.provider,
+            target.settings.model,
+        )
+        if self._screen_mounted_accessor():
+            self._sync_console_chat_core_state()
+            self._sync_console_settings_summary()
+            self._sync_console_control_bar()
+        if not fence_matches(expected_active_id=intent.session_id):
+            return rollback_and_release(
+                "Console changed before the first chat finished opening. It will retry.",
+            )
+        try:
+            acknowledged = run_if_runtime_config_generation_current(
+                intent.config_revision,
+                lambda: self.app_instance.pending_handoffs.acknowledge_current(claim),
+            )
+        except Exception as exc:  # noqa: BLE001 - mount/resume must not fail
+            self._log_first_chat_handoff_exception("guarded-acknowledgement", exc)
+            return rollback_and_release(
+                "The first chat could not be acknowledged yet. It will retry.",
+            )
+        if not acknowledged:
+            return rollback_and_release(
+                "The first chat could not be acknowledged yet. It will retry.",
+            )
+        self._first_chat_handoff_notified_revision = None
+        return True
 
     # -- Session-local character reactions ----------------------------------
